separate location updates from general transaction updates
